+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - build: use official bindings instead of manual jar import
 - build: propagate versions to nuspec
 - build: centralize tfm versions
+- ios: fix ios trusted roots
 - Microsoft.Extensions.* 10.0.4
 - Square.OkHttp3.Android 5.3.2.2
 - Square.OkIO 3.16.4.2